+++ tests/integration_tests/test_srnn_hjh.py
@@ -1,8 +1,8 @@
 import pytorch_lightning as pl
 
-from src.datamodule.speech_command_datamodule import SpeechCommandDataModule
-from src.model.lit_wrapper import LitWrapper
-from src.model.srnn_hjh import SrnnHJH
+from snn_voice.datamodule.speech_command_datamodule import SpeechCommandDataModule
+from snn_voice.model.lit_wrapper import LitWrapper
+from snn_voice.model.srnn_hjh import SrnnHJH
 from tests.evaluate import evaluate_model
